TST/CLN: Remove groupby.test_allowlist
